# docs/swagger.yaml
definitions:
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controllers.LikeResponse:
    properties:
      likes_count:
        example: 42
        type: integer
      message:
        example: Post liked successfully
        type: string
    type: object
...
  controllers.MessageResponse:
    properties:
      message:
        type: string
    type: object
  controllers.PaginationMetadata:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  controllers.PostInput:
    properties:
      content:
        type: string
      image_urls:
        items:
          type: string
        type: array
    type: object
  controllers.PostListResponse:
    properties:
      metadata:
        $ref: '#/definitions/controllers.PaginationMetadata'
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
    type: object
...
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      parent_id:
        type: integer
      post_id:
        type: integer
      replies:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Like:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Post:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      i_liked:
        type: boolean
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      likes:
        items:
          $ref: '#/definitions/models.Like'
        type: array
      likes_count:
        type: integer
      post_type:
        type: string
      shares_count:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      views_count:
        type: integer
    type: object
  models.User:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      followers:
        description: Self-referential many-to-many for followers and following.
        items:
          $ref: '#/definitions/models.User'
        type: array
      following:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        type: integer
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      profile_picture:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /posts:
    get:
      description: Get a list of all posts with optional pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Posts per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.PostListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: List all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new post with content and optional image URL
      parameters:
      - description: Post Input
        in: body
        name: postInput
        required: true
        schema:
          $ref: '#/definitions/controllers.PostInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete a post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Edit a post's content and image URL
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Post Input
        in: body
        name: postInput
        required: true
        schema:
          $ref: '#/definitions/controllers.PostInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Edit a post
      tags:
      - posts
...
  /posts/{id}/like:
    delete:
      consumes:
      - application/json
      description: Remove a like from a post by the authenticated user
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully unliked the post
          schema:
            $ref: '#/definitions/controllers.LikeResponse'
        "400":
          description: Invalid post ID or Like not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unlike a post
      tags:
      - likes
    post:
      consumes:
      - application/json
      description: Add a like to a post by the authenticated user
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully liked the post
          schema:
            $ref: '#/definitions/controllers.LikeResponse'
        "400":
          description: Invalid post ID or Already liked
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Post not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like a post
      tags:
      - likes
  /posts/timeline:
    get:
      description: Get posts created by the authenticated user and those they follow
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get timeline posts
      tags:
      - posts
...
swagger: "2.0"